Add files via upload
This a is PCSX2 .pnach code (written by Alpha) for use with Biohazard Outbreak online scenarios which have had the enemies types and numbers modded in the .emd files. It is necessary because otherwise the enemies will only show up the first game that is played after booting the game. It is not known exactly why this happens but appears to be because the enemies are somehow tied to the item sets. The first item set is always the default item set used in the  first scenario after the game is loaded. Then  the item set changes randomly in subsequent scenarios. Alpha found a way to ensure the first (default) item set is ALWAYS loaded and wrote this .pnach for this purpose.

To use this file place it in the cheats folder in your PCSX2 folder and check Enable Cheats in the System tab in PCSX2.